fix snapshot conflict
